requirements.txt
# AI Research Portal — Phases 2 & 3 (RAG + Agentic Pipeline)
# Pin major+minor to ensure reproducibility.

# --- Phase 1 (carried forward) ---
mlx-lm>=0.0.1
pdfplumber>=0.10.0
python-dotenv>=1.0.0
transformers>=4.46.0

# --- Phase 2: RAG pipeline ---
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
numpy>=1.24.0
tiktoken>=0.5.0
requests>=2.31.0

# --- Phase 3: Agentic pipeline (optional cloud LLM providers) ---
# Install these only if using --provider openai or --provider anthropic.
# The default MLX provider requires no additional packages.
openai>=1.0.0
anthropic>=0.18.0